Remove the need to wait for the EndRx event in the satellite before forwarding
  Due to interference computations necessary at the StartRx event on ground
entities, satellite carriers (satellite-phy-rx-carrier-uplink.cc) had to wait
the end of the reception (EndRxData) before emiting a packet back, substracting
the time it took to receive it to simulate instant forwarding.
  This lead to issues regarding messages whose duration was longer than
propagation time as it would mean scheduling receive event on ground entities
somewhere in the past of the simulation.